fix(a2-audit): Paso 3 - migrar a integer 41 casos en Cashier, Printers y Fiscal
FASE A2 - PASO 3: Migración de 143 casts (float) residuales.

CAMBIOS EN CÓDIGO (41 cambios en 8 archivos):

📦 CASHIER (16 cambios):
   - TipPayoutResource: amount como int
   - CashMovementController: amount como int (3 ocurrencias)
   - CashierReportController:
     * Breakdown por método (cash/card/transfer/gift_card): amount + tips
     * Movement summary (withdrawals/deposits/adjustments)
     * lastCount->total_counted como int

📦 PRINTERS (12 cambios):
   - ReceiptFormatter: price, subtotal, tax, discount, total como int
   - PrintReceiptOnOrderPaid:
     * Items: price + subtotal como int
     * Order: subtotal, tax, discount, total como int

📦 FISCAL (13 cambios):
   - DteDocumentResource: net_amount, tax_amount, exempt_amount, total_amount
   - SalesBookController: totales generales y por tipo
   - DteXmlGenerator: eliminado cast redundante (entidad ya cast a int)

NO TOCADOS (float legítimo, 8 casos):
   - TipPolicyController (4): waiter_percentage, pool_percentage
   - PrintReceiptOnOrderPaid (2): tax_rate_snapshot
   - DteType (2): tax_rate_snapshot

PROGRESO ACUMULADO FASE A2:
✅ Paso 1: 13 casos
✅ Paso 2: 18 casos
✅ Paso 3: 41 casos
📊 Total: 72/143 casos migrados (~50%)

RESULTADO:
✅ 1021 tests passing, 0 failed
✅ 3050 assertions
✅ Consistencia con ADR-018

PRÓXIMO PASO: Paso 4 (Inventory + Recipes, ~59 casos)

+++ app/Modules/Cashier/Interfaces/Controllers/CashMovementController.php
@@ -72,7 +72,7 @@ class CashMovementController extends Controller
                 MovementType::WITHDRAWAL => $this->movementService->withdrawal(
                     $session,
                     $user,
-                    (float) $validated['amount'],
+                    (int) $validated['amount'],  // ADR-018
                     $validated['reason'],
                     $validated['notes'] ?? null,
                     $authorizer
@@ -80,14 +80,14 @@ class CashMovementController extends Controller
                 MovementType::DEPOSIT => $this->movementService->deposit(
                     $session,
                     $user,
-                    (float) $validated['amount'],
+                    (int) $validated['amount'],  // ADR-018
                     $validated['reason'],
                     $validated['notes'] ?? null
                 ),
                 MovementType::ADJUSTMENT => $this->movementService->adjustment(
                     $session,
                     $user,
-                    (float) $validated['amount'],
+                    (int) $validated['amount'],  // ADR-018
                     $validated['reason'],
                     $authorizer,
                     $validated['notes'] ?? null
